API/requirements.txt
asgiref==3.3.1
astroid==2.7.3
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
distlib==0.3.3
Django==3.1.4
django-cors-headers==3.6.0
django-extensions==3.1.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-tqdm==1.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
filelock==3.3.1
idna==2.10
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.3
mccabe==0.6.1
pipenv==2021.5.29
platformdirs==2.3.0
PyJWT==2.3.0
pylint==2.10.2
pytz==2020.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tqdm==4.54.1
urllib3==1.26.5
virtualenv==20.9.0
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
